requirements.txt
llama-cpp-python
pandas
tqdm
numpy
